pyproject.toml
[project]
name = "cordfeeder"
version = "0.1.0"
description = "RSS/Atom feed reader that posts to Discord channels"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9,<4",
    "aiosqlite>=0.20,<1",
    "discord-py>=2.4,<3",
    "feedparser>=6.0,<7",
    "python-dateutil>=2.9,<3",
    "python-dotenv>=1.0,<2",
]

[project.scripts]
cordfeeder = "cordfeeder.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B"]
ignore = [
    "S101",   # assert usage is fine
    "S106",   # hardcoded passwords in test fixtures
    "S311",   # random for jitter, not crypto
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]  # test strings/XML can exceed line length

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["feedparser", "feedparser.*"]
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.15.4",
    "types-python-dateutil>=2.9.0.20260124",
]
